requirements.txt
opencv-python~=4.10.0.84
numpy~=2.2.1
requests~=2.32.3
openpyxl~=3.1.5
PyYAML~=6.0.2
pillow~=11.1.0
PyQt5~=5.15.11